pyproject.toml
[project]
name = "bharatagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain[google-genai]>=1.0.3",
    "langchain-community>=0.4.1",
    "langgraph>=1.0.2",
    "python-dotenv>=1.2.1",
    "duckdb-engine>=0.17.0",
    "grandalf>=0.8",
]
